Split webviewMessageHandler out of ClineProvider.ts
